Let Dyninst itself be compiled with Clang on Linux
With clang-3.3-0.10.rc3.fc19.x86_64 and libstdc++-4.8.2-1.fc19.x86_64,
the fixes required were as follows:

- cmake files looking for CMAKE_COMPILER_IS_GNUCXX need to also
  recognize CMAKE_CXX_COMPILER_ID "Clang".

- The ListHash macro had a pointer->int cast that halted Clang.

- Several templates in rose/*InstructionSemantics.h couldn't be resolved
  implicitly; they're now either made explicit or converted to typename.

- ppc32::cr0.val() is apparently not const enough for Clang as a
  template parameter; changed to ppc32::icr0.

+++ common/src/List.h
@@ -46,7 +46,7 @@
 #define DO_INLINE_F
 #endif
 
-#define ListHash(ptr, size) (((unsigned int)(ptr) % (unsigned int)(size)))
+#define ListHash(ptr, size) (((unsigned int)(intptr_t)(ptr) % (unsigned int)(size)))
 
 template <class DataType> class List;
 template <class DataType, class KeyType> class ListBase;
